Move to skill tree view
