+++ src/intermission.rs
@@ -1,5 +1,6 @@
 mod character_scene;
 mod reward_scene;
+mod skilltree_view;
 
 pub use character_scene::*;
 pub use reward_scene::*;
